refactor(core): Introduce `@n8n/constants` package
