+++ packages/@n8n/constants/.eslintrc.js
@@ -1,0 +1,7 @@
+const sharedOptions = require('@n8n/eslint-config/shared');
+
+/** @type {import('@types/eslint').ESLint.ConfigData} */
+module.exports = {
+	extends: ['@n8n/eslint-config/base'],
+	...sharedOptions(__dirname),
+};
