+++ packages/@n8n/constants/jest.config.js
@@ -1,0 +1,7 @@
+/** @type {import('jest').Config} */
+module.exports = {
+	...require('../../../jest.config'),
+	transform: {
+		'^.+\\.ts$': ['ts-jest', { isolatedModules: false }],
+	},
+};
