+++ packages/cli/package.json
@@ -92,6 +92,7 @@
     "@n8n/api-types": "workspace:*",
     "@n8n/client-oauth2": "workspace:*",
     "@n8n/config": "workspace:*",
+    "@n8n/constants": "workspace:^",
     "@n8n/di": "workspace:*",
     "@n8n/localtunnel": "3.0.0",
     "@n8n/n8n-nodes-langchain": "workspace:*",
