+++ packages/cli/src/constants.ts
@@ -72,50 +72,6 @@ export const WORKFLOW_REACTIVATE_MAX_TIMEOUT = 24 * 60 * 60 * 1000; // 1 day
 
 export const SETTINGS_LICENSE_CERT_KEY = 'license.cert';
 
-export const LICENSE_FEATURES = {
-	SHARING: 'feat:sharing',
-	LDAP: 'feat:ldap',
-	SAML: 'feat:saml',
-	LOG_STREAMING: 'feat:logStreaming',
-	ADVANCED_EXECUTION_FILTERS: 'feat:advancedExecutionFilters',
-	VARIABLES: 'feat:variables',
-	SOURCE_CONTROL: 'feat:sourceControl',
-	API_DISABLED: 'feat:apiDisabled',
-	EXTERNAL_SECRETS: 'feat:externalSecrets',
-	SHOW_NON_PROD_BANNER: 'feat:showNonProdBanner',
-	WORKFLOW_HISTORY: 'feat:workflowHistory',
-	DEBUG_IN_EDITOR: 'feat:debugInEditor',
-	BINARY_DATA_S3: 'feat:binaryDataS3',
-	MULTIPLE_MAIN_INSTANCES: 'feat:multipleMainInstances',
-	WORKER_VIEW: 'feat:workerView',
-	ADVANCED_PERMISSIONS: 'feat:advancedPermissions',
-	PROJECT_ROLE_ADMIN: 'feat:projectRole:admin',
-	PROJECT_ROLE_EDITOR: 'feat:projectRole:editor',
-	PROJECT_ROLE_VIEWER: 'feat:projectRole:viewer',
-	AI_ASSISTANT: 'feat:aiAssistant',
-	ASK_AI: 'feat:askAi',
-	COMMUNITY_NODES_CUSTOM_REGISTRY: 'feat:communityNodes:customRegistry',
-	AI_CREDITS: 'feat:aiCredits',
-	FOLDERS: 'feat:folders',
-	INSIGHTS_VIEW_SUMMARY: 'feat:insights:viewSummary',
-	INSIGHTS_VIEW_DASHBOARD: 'feat:insights:viewDashboard',
-	INSIGHTS_VIEW_HOURLY_DATA: 'feat:insights:viewHourlyData',
-	API_KEY_SCOPES: 'feat:apiKeyScopes',
-} as const;
-
-export const LICENSE_QUOTAS = {
-	TRIGGER_LIMIT: 'quota:activeWorkflows',
-	VARIABLES_LIMIT: 'quota:maxVariables',
-	USERS_LIMIT: 'quota:users',
-	WORKFLOW_HISTORY_PRUNE_LIMIT: 'quota:workflowHistoryPrune',
-	TEAM_PROJECT_LIMIT: 'quota:maxTeamProjects',
-	AI_CREDITS: 'quota:aiCredits',
-	INSIGHTS_MAX_HISTORY_DAYS: 'quota:insights:maxHistoryDays',
-	INSIGHTS_RETENTION_MAX_AGE_DAYS: 'quota:insights:retention:maxAgeDays',
-	INSIGHTS_RETENTION_PRUNE_INTERVAL_DAYS: 'quota:insights:retention:pruneIntervalDays',
-} as const;
-export const UNLIMITED_LICENSE_QUOTA = -1;
-
 export const CREDENTIAL_BLANKING_VALUE = '__n8n_BLANK_VALUE_e5362baf-c777-4d57-a609-6eaf1f9e87f6';
 
 export const UM_FIX_INSTRUCTION =
