+++ packages/cli/src/decorators/controller.registry.ts
@@ -1,4 +1,5 @@
 import { GlobalConfig } from '@n8n/config';
+import type { BooleanLicenseFeature } from '@n8n/constants';
 import { Container, Service } from '@n8n/di';
 import { Router } from 'express';
 import type { Application, Request, Response, RequestHandler } from 'express';
@@ -9,7 +10,6 @@ import type { ZodClass } from 'zod-class';
 import { AuthService } from '@/auth/auth.service';
 import { inProduction, RESPONSE_ERROR_MESSAGES } from '@/constants';
 import { UnauthenticatedError } from '@/errors/response-errors/unauthenticated.error';
-import type { BooleanLicenseFeature } from '@/interfaces';
 import { License } from '@/license';
 import { userHasScopes } from '@/permissions.ee/check-access';
 import type { AuthenticatedRequest } from '@/requests';
