+++ packages/cli/src/services/community-packages.service.ts
@@ -1,4 +1,5 @@
 import { GlobalConfig } from '@n8n/config';
+import { LICENSE_FEATURES } from '@n8n/constants';
 import { Service } from '@n8n/di';
 import axios from 'axios';
 import { exec } from 'child_process';
@@ -9,7 +10,6 @@ import { UnexpectedError, UserError, type PublicInstalledPackage } from 'n8n-wor
 import { promisify } from 'util';
 
 import {
-	LICENSE_FEATURES,
 	NODE_PACKAGE_PREFIX,
 	NPM_COMMAND_TOKENS,
 	NPM_PACKAGE_STATUS_GOOD,
