+++ turbo.json
@@ -28,6 +28,7 @@
 			"dependsOn": [
 				"@n8n/api-types#lint",
 				"@n8n/config#lint",
+				"@n8n/constants#lint",
 				"@n8n/di#lint",
 				"@n8n/client-oauth2#lint",
 				"@n8n/imap#lint",
